% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air-data.R
\docType{data}
\name{air}
\alias{air}
\title{Extract from AgrImOnIA dataset}
\format{
An object of class \code{data.frame}.
}
\source{
\href{https://zenodo.org/record/6620530#.Y1ZsD-xBz1I}{link}
}
\usage{
data(air)
}
\description{
TO DO.
}
\keyword{dataset}
